+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 
 * Savestates now restore better file descriptors
 * Fix SDL getting window dimensions of zero
+* Better handling of opendir/readdir and savefiles
+* Fix the Unity job handling not calling the completed callback
 
 ## [1.4.7] - 2025-10-06
 ### Added